DRY tests into utils
